Break out get_list_from_csv again
This appears to work now as a result of the import refactor on the Talon side.
